feat!: Change FON data structure to queue of sets
A queue is now store the history of the First-Order-Neighborhoods. The
set to return is now sought leaping from top to bottom until a non-empty
set is found. That is: If the current (top of the queue) is empty, go
and see the bottom. If that is empty as well, go see the second-to-top.
If that is empty, go and see the second-to-bottom. And so on..

Switching from top to bottom allows to keep the growth of the cell on a
linear path front-advancing path and should avoid ending up with cells
shaped like 'T's