Refactoring to add another agglomerator
A big refactoring has been done to take into account an agglomerator
that is full Front Advancing and an agglomerator that is solely
biconnected as the last one. Probably there is a way to do the thing a
bit more properly (like to have the biconnected as a further child) but
right now it is should be ok. Further improvement has been the new
argument to take into account the type of Isotropic Agglomerator.