chore: Adapt examples to new version
So far, most of the time, we pass to CoMMA directly the `numpy` array of
the centers coming from `dualGPy`. It seems to work fine, meaning that
C++ see the right dimensions and the right data. However, I don't know
if this is the safest way to do it (possibly as a list of list?).

WARNING: I was not able to reconstruct the centers for the mesh in
`test_from_comma_sources.py` hence it won't work with the new version

Maybe @albiremo has some ideas about those two items...