feat!: Add priority weights to decide the order of the agglomeration
This is a huge commit. It aims at having a deterministic order of
agglomeration which depends only on some user defined weights, a new
input. To do that we replace in strategic points `unordered_map`/`set`
(which have unpredictable behavior due to unordered nature) with
`priority_queue` and standard (ordered) `set`.
Moreover, a thorough cleaning.

WARNING: it compiles without errors (and warnings) but tests have not yet
been evaluated