feat!: Bring back gradual seed pool
Instead of building all the seed queues at the beginning, only the one
related to the highest cell type (usually corners in 3D and ridges in
2D) is built. Then, the others are gradually filled with the neighbors
of the coarse cells already agglomerated.

This allows to have such neighbors at the head of the queue so that they
are chosen with priority and therefore, coarse cells grows just one next
to the other without leaving any hole between them.

This is a partial commit since the update of the seed pool is done only
with the first order neighborhood remaining at the end which is only a
subset of the neighbors of the coarse cell.