refactor(neigh): Do not sort candidates after every update
We used to add the new neighbours to the candidates and then sort them
by using the weights. Now, we just sort the new neighbours and add them
at the back of the candidates queue. This leaves candidates with lower
neighborhood order wrt to the original seed on top of the queue.

It addresses one item of #32