feat!(agglo): Add iterative version
A new input argument the number of iterations for the choice of the best
fine cell to add. It is defaulted to 1. It works like this:
- If = 1, as usual
- If = n, the algorithm to choose the next best fine cell tries to
agglomerate n times. The direct neighbours which yields the best CC
after n agglomeration step is retained.

Shared faces have highest priority.

The process is implemented with a recursive function. It might need some
optimization.

Right now the structure is quite chaotic but it works