tests: Add simple test for iterative FC choice
The mesh has been created so that an I-shaped CC leads to a higher AR
than an L-shaped one. However, if the seed is a corner, the additional
check in choose_best_fc regarding the neighbourhood order, lead us to
prefer the L-shaped cell. We start here from the bottom middle so that
the above check is not strict and let the algorithm choose the cell
according to the AR.

The standard algorithm will thus pick the I-shaped CC (leading to a
T-shaped final CC). The iterative, on the contrary, will pick the
L-shaped CC (leading to a square final CC)