feat!: Add dedicated class for computing AR
AR is used lightly, and in general, it is just a quantity that one
wishes to minimize. Several (hopefully interesting) ways are available
as children of the abstract base class.

For the moment, it is not used. Indeed, it should be tested. Trivial
tests (see `// DEBUG` in code) have been performed and passed.